Known issues:
- strange behavior on create momospace styles text
- no idea how to create code block (maybe momospace?)

+++ TMessagesProj/src/main/java/org/telegram/messenger/ImageReceiver.java
@@ -2309,7 +2309,7 @@ public class ImageReceiver implements NotificationCenter.NotificationCenterDeleg
         }
     }
 
-    public void setImageX(int x) {
+    public void setImageX(float x) {
         imageX = x;
     }
